remove post processing cause it no work

+++ Assets/materials/composite.shader
@@ -51,14 +51,14 @@
 			float2 cameraSize;
 			float cameraAngle;
 
-            fixed4 frag (v2f i) : SV_Target
+            float4 frag (v2f i) : SV_Target
             {
 				#if UNITY_UV_STARTS_AT_TOP
 				if (_MainTex_TexelSize.y < 0)
 					i.uv.y = 1 - i.uv.y;
 				#endif
 
-                fixed4 col = tex2D(_MainTex, i.uv);
+                float4 col = tex2D(_MainTex, i.uv);
 				fixed viewMask = tex2D(_ViewMask, i.uv);
 
 				float2 ratio = cameraSize / FOWsize;
@@ -81,6 +81,7 @@
 				uv /= (FOWsize * 2);
 
 				fixed fow = tex2D(_FogOfWar, uv)*_FogOfWarBrightness;
+
 				return col * max(viewMask.rrrr, fow.rrrr);
             }
             ENDCG
